Add the second recipe page
Add an image of the second dish with img element.
Add an “Ingredients” heading followed by an unordered list of the ingredients needed for the second recipe.
Add a “Steps” heading followed by an ordered list of the steps needed for making the dish.
Add and ordered list in index with the names of the recipes to go to the page.